🚚 move : Move lib, components folder outside of app folder
