+++ src/app/hotspringList/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React, { useState } from "react";
-import { useGetHotspringBySidoQuery } from "../lib/hotspring/hotspringApi";
+import { useGetHotspringBySidoQuery } from "@/lib/hotspring/hotspringApi";
 
 
 const HotspringListPage = () => {
